requirements.txt
python-dotenv==1.2.1
requests==2.32.5
toml==0.10.2